+++ include/sys/dbg/cmd/ls.h
@@ -29,6 +29,6 @@
  * @param argv the arguments
  * @return 0 on success
  */
-s32 cons_cmd_ls(s32 argc,char **argv);
+int cons_cmd_ls(size_t argc,char **argv);
 
 #endif /* LS_H_ */
